Move toTrinoDictionaryPage to ParquetTestUtils
